refactor: change the rest of the Navigation following the Tooltip support
