+++ src/Navigation/components/shared/disableHoverEvents.ts
@@ -1,0 +1,7 @@
+import { PointerEvent } from "react";
+
+export const disableHoverEvents = {
+  onPointerMove: (e: PointerEvent<Element>): void => e.preventDefault(),
+  onPointerEnter: (e: PointerEvent<Element>): void => e.preventDefault(),
+  onPointerLeave: (e: PointerEvent<Element>): void => e.preventDefault(),
+};
